Add files via upload
These two files were used to implement both the raspberry pi and arduino uno together using serial communication. This code allows for the camera to detect the quadrant that an aruco marker falls in and then sends that information from the pi to the arduino. The arduino then determines the desired angle for the motor and sends out the desired and actual position of the motor using an encoder library back to the pi. This is the read in through the pi and displayed on the LCD screen.